assetalpha_v1/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic[email]==2.9.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
yfinance==0.2.44
pandas==2.2.3
numpy==1.26.4
scipy==1.14.1
alembic==1.13.3
python-multipart==0.0.12
slowapi==0.1.9
